+++ CHANGES.rst
@@ -1,7 +1,11 @@
 0.17.2 (unreleased)
 ===================
 
--
+- Update minimum version of astropy to 5.3.0 in order to fix a bug due to a breaking
+  change in astropy. [#258]
+
+- Update minimum version of numpy to 1.22 as this is the oldest version of numpy
+  which is currently supported. [#258]
 
 0.17.1 (2023-08-03)
 ===================