+++ CHANGES.rst
@@ -7,13 +7,14 @@
 - Update minimum version of numpy to 1.22 as this is the oldest version of numpy
   which is currently supported. [#258]
 
+
+- Allow DNode and LNode subclass instances to be assigned to tree attributes (without
+  immediate validation). [#527]
+
 0.17.1 (2023-08-03)
 ===================
 
 - Fix newly required units from rand [#256]
-
-- Allow DNode and LNode subclass instances to be assigned to tree attributes (without
-  immediate validation). [#527]
 
 - Add checks for for association processing [#241]