+++ CHANGES.rst
@@ -12,6 +12,8 @@
 
 - Fix the initialization of empty DataModels and clean up the datamodel core. [#251]
 
+- Add slope and error to dark RefModel and tests. [#280]
+
 0.17.1 (2023-08-03)
 ===================